requirements.txt
psutil==5.8.0
qrcode==6.1
Pillow==8.2.0
